+++ src/pages/PlayList/DetailList/BannerBtns.module.css
@@ -1,0 +1,89 @@
+.banner_btns {
+  width: 100%;
+  display: flex;
+  gap: 8px;
+}
+
+.play_btn {
+  flex: 1;
+  border-radius: 18px;
+  border: none;
+  overflow: hidden;
+
+  height: 36px;
+  font-size: 14px;
+
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: 8px;
+  font-weight: 500;
+  cursor: pointer;
+}
+
+.play_btn:hover {
+  background-color: rgba(255, 255, 255, 0.8);
+}
+
+.icons_box {
+  flex: 1;
+  display: flex;
+  gap: 8px;
+}
+
+.icon_btn {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+
+  width: 36px;
+  height: 36px;
+
+  background-color: rgba(255, 255, 255, 0.2);
+  color: white;
+  border: none;
+  border-radius: 50%;
+  cursor: pointer;
+}
+
+.icon_btn:hover {
+  background-color: rgba(255, 255, 255, 0.3);
+}
+
+.menu_list {
+  list-style: none;
+  padding: 0;
+  margin: 0;
+  display: flex;
+  flex-direction: column;
+  width: 300px;
+  background-color: white;
+  color: black;
+
+  border-radius: 12px;
+  overflow: hidden;
+  box-shadow: rgba(0, 0, 0, 0.1) 0px 4px 32px 0px;
+}
+
+.menu_item {
+  display: flex;
+  gap: 12px;
+  font-size: 14px;
+  height: 40px;
+  align-items: center;
+  padding: 0 12px;
+  cursor: pointer;
+}
+
+.icon_center {
+  width: 24px;
+  height: 24px;
+
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.menu_item:hover {
+  background-color: #f5f5f5;
+}
