+++ src/pages/PlayList/Like/Banner.module.css
@@ -33,12 +33,6 @@
 .banner_title {
   font-size: 28px;
   font-weight: 700;
-
-  display: -webkit-box;
-  -webkit-line-clamp: 2;
-  -webkit-box-orient: vertical;
-  overflow: hidden;
-  text-overflow: ellipsis;
 }
 
 .banner_meta {
@@ -46,30 +40,12 @@
   flex-direction: column;
   gap: 8px;
 
-  font-size: 12px;
-}
-
-.channel_info {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-}
-
-.channel_img {
-  width: 24px;
-  height: 24px;
-  border-radius: 50%;
+  font-size: 14px;
 }
 
 .banner_detail {
-  display: flex;
-  align-items: center;
-  gap: 4px;
+  font-size: 12px;
   opacity: 0.7;
-}
-
-.banner_detail > p {
-  margin: 0;
 }
 
 .banner_btns {
@@ -78,7 +54,7 @@
   gap: 8px;
 }
 
-.play_btn {
+.banner_btns > button {
   flex: 1;
   border-radius: 18px;
   border: none;
@@ -95,32 +71,16 @@
   cursor: pointer;
 }
 
-.play_btn:hover {
+.like_btn:hover {
   background-color: rgba(255, 255, 255, 0.8);
 }
 
-.icons_box {
-  flex: 1;
-  display: flex;
-  gap: 8px;
-}
-
-.icons_box > button {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-
-  width: 36px;
-  height: 36px;
-
+.shuffle_btn {
   background-color: rgba(255, 255, 255, 0.2);
   color: white;
-  border: none;
-  border-radius: 50%;
-  cursor: pointer;
 }
 
-.icons_box > button:hover {
+.shuffle_btn:hover {
   background-color: rgba(255, 255, 255, 0.3);
 }
